requirements.txt
streamlit
langchain-groq
langchain
langchain-community
pypdf
sentence-transformers
faiss-cpu
python-dotenv